% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Create API}
\usage{
api(file, preroute = pre, postroute = post)
}
\arguments{
\item{file}{Character. Path to api spec.}

\item{preroute}{Function. Preroute handler.}

\item{postroute}{Functino. Postroute handler.}
}
\value{
A plumber API object.
}
\description{
Create the API object
}
