% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_geo_convert.R
\name{finbif_geo_convert}
\alias{finbif_geo_convert}
\title{FinBIF geo-conversion}
\usage{
finbif_geo_convert(
  input,
  output = "none",
  geo = c("point", "bbox", "footprint"),
  crs = "wgs84",
  ...
)
}
\arguments{
\item{input}{Character or Integer. Either the path to a Zip archive or
tabular data file that has been downloaded from "laji.fi", a URI
linking to such a data file (e.g.,
\url{https://tun.fi/HBF.53254}) or an integer
representing the URI (i.e., \code{53254}).}

\item{output}{Character. Output file format in the form of a file extension.
See \code{show_formats()} for a list of available file formats. Use \code{"none"}
(default) to prevent writing output to a file.}

\item{geo}{Character. Geometry of output. One of \code{"point"}, \code{"bbox"} or "
\code{"footprint"}.}

\item{crs}{Character or Integer. Coordinate reference system of output. One
of \code{"euref"}, \code{"ykj"}, \code{"wgs84"} or integer indicating an EPSG code.}

\item{...}{Other options passed to \code{finbif::finbif_occurrence_load}.}
}
\value{
An \code{{sf}} package simple feature object (invisibly). And if
\code{output != "none"} then the spatial data file(s) will be written to the
current working directory.
}
\description{
Convert FinBIF data to geographic formats.
}
