% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise_id.R
\name{sanitise_id}
\alias{sanitise_id}
\title{Sanitise input ID}
\usage{
sanitise_id(x)
}
\arguments{
\item{x}{A FinBIF download identifier}
}
\value{
A sanitised FinBIF download identifier
}
\description{
Convert FinBIF download identifier inputs of different types
}
